libcamera: base: utils: Prevent hex signed extension
When converting signed values to hex strings using the utils::hex
helpers, signed negative values get sign extended to the larger 64 bit
type.

This causes values such as 0x80 to be printed as 0xffffffffffffff80
instead.

Resolve this by first converting all signed integer types to an unsigned
type of the same size before extending to the larger type.

+++ include/libcamera/base/utils.h
@@ -95,7 +95,7 @@ _hex hex(T value, unsigned int width = 0);
 template<>
 inline _hex hex<int8_t>(int8_t value, unsigned int width)
 {
-	return { static_cast<uint64_t>(value), width ? width : 2 };
+	return { static_cast<uint64_t>(static_cast<uint8_t>(value)), width ? width : 2 };
 }
 
 template<>
@@ -107,7 +107,7 @@ inline _hex hex<uint8_t>(uint8_t value, unsigned int width)
 template<>
 inline _hex hex<int16_t>(int16_t value, unsigned int width)
 {
-	return { static_cast<uint64_t>(value), width ? width : 4 };
+	return { static_cast<uint64_t>(static_cast<uint16_t>(value)), width ? width : 4 };
 }
 
 template<>
@@ -119,7 +119,7 @@ inline _hex hex<uint16_t>(uint16_t value, unsigned int width)
 template<>
 inline _hex hex<int32_t>(int32_t value, unsigned int width)
 {
-	return { static_cast<uint64_t>(value), width ? width : 8 };
+	return { static_cast<uint64_t>(static_cast<uint32_t>(value)), width ? width : 8 };
 }
 
 template<>
